refactor(services): extract shared wasm spec matcher into wasmspec
