+++ internal/services/sep41/validator.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 
 	"github.com/alitto/pond/v2"
-	set "github.com/deckarep/golang-set/v2"
 	"github.com/jackc/pgx/v5"
 	"github.com/stellar/go-stellar-sdk/strkey"
 	"github.com/stellar/go-stellar-sdk/support/log"
@@ -15,6 +14,7 @@ import (
 	"github.com/stellar/wallet-backend/internal/data"
 	"github.com/stellar/wallet-backend/internal/indexer/types"
 	"github.com/stellar/wallet-backend/internal/services"
+	"github.com/stellar/wallet-backend/internal/services/wasmspec"
 )
 
 const ProtocolID = "SEP41"
@@ -254,127 +254,88 @@ func decodeContractAddr(hash types.HashBytea) (string, bool) {
 	return addr, true
 }
 
-// scSpecTypeNames maps XDR ScSpecType values to human-readable type names used
-// during signature validation.
-var scSpecTypeNames = map[xdr.ScSpecType]string{
-	xdr.ScSpecTypeScSpecTypeBool:         "bool",
-	xdr.ScSpecTypeScSpecTypeU32:          "u32",
-	xdr.ScSpecTypeScSpecTypeI32:          "i32",
-	xdr.ScSpecTypeScSpecTypeU64:          "u64",
-	xdr.ScSpecTypeScSpecTypeI64:          "i64",
-	xdr.ScSpecTypeScSpecTypeU128:         "u128",
-	xdr.ScSpecTypeScSpecTypeI128:         "i128",
-	xdr.ScSpecTypeScSpecTypeU256:         "u256",
-	xdr.ScSpecTypeScSpecTypeI256:         "i256",
-	xdr.ScSpecTypeScSpecTypeAddress:      "Address",
-	xdr.ScSpecTypeScSpecTypeMuxedAddress: "MuxedAddress",
-	xdr.ScSpecTypeScSpecTypeString:       "String",
-	xdr.ScSpecTypeScSpecTypeBytes:        "Bytes",
-	xdr.ScSpecTypeScSpecTypeSymbol:       "Symbol",
-	xdr.ScSpecTypeScSpecTypeVec:          "Vec",
-	xdr.ScSpecTypeScSpecTypeMap:          "Map",
-	xdr.ScSpecTypeScSpecTypeOption:       "Option",
-	xdr.ScSpecTypeScSpecTypeResult:       "Result",
-	xdr.ScSpecTypeScSpecTypeTuple:        "Tuple",
-	xdr.ScSpecTypeScSpecTypeBytesN:       "BytesN",
-	xdr.ScSpecTypeScSpecTypeUdt:          "UDT",
-	xdr.ScSpecTypeScSpecTypeVoid:         "void",
-}
-
-// functionSpec defines the expected signature for a SEP-41 token function.
-type functionSpec struct {
-	name            string
-	expectedInputs  []contractFunctionInputSpec
-	expectedOutputs []string
-}
-
-type contractFunctionInputSpec struct {
-	name     string
-	typeName string
-}
-
 // requiredFunctions defines all required functions for SEP-41 token standard
 // compliance. A contract must implement all of these functions with the exact
 // signatures specified.
-var requiredFunctions = []functionSpec{
+var requiredFunctions = []wasmspec.FunctionSpec{
 	{
-		name:            "balance",
-		expectedInputs:  []contractFunctionInputSpec{{name: "id", typeName: "Address"}},
-		expectedOutputs: []string{"i128"},
+		Name:            "balance",
+		ExpectedInputs:  []wasmspec.InputSpec{{Name: "id", TypeName: "Address"}},
+		ExpectedOutputs: []string{"i128"},
 	},
 	{
-		name:            "allowance",
-		expectedInputs:  []contractFunctionInputSpec{{name: "from", typeName: "Address"}, {name: "spender", typeName: "Address"}},
-		expectedOutputs: []string{"i128"},
+		Name:            "allowance",
+		ExpectedInputs:  []wasmspec.InputSpec{{Name: "from", TypeName: "Address"}, {Name: "spender", TypeName: "Address"}},
+		ExpectedOutputs: []string{"i128"},
 	},
 	{
-		name:            "decimals",
-		expectedInputs:  []contractFunctionInputSpec{},
-		expectedOutputs: []string{"u32"},
+		Name:            "decimals",
+		ExpectedInputs:  []wasmspec.InputSpec{},
+		ExpectedOutputs: []string{"u32"},
 	},
 	{
-		name:            "name",
-		expectedInputs:  []contractFunctionInputSpec{},
-		expectedOutputs: []string{"String"},
+		Name:            "name",
+		ExpectedInputs:  []wasmspec.InputSpec{},
+		ExpectedOutputs: []string{"String"},
 	},
 	{
-		name:            "symbol",
-		expectedInputs:  []contractFunctionInputSpec{},
-		expectedOutputs: []string{"String"},
+		Name:            "symbol",
+		ExpectedInputs:  []wasmspec.InputSpec{},
+		ExpectedOutputs: []string{"String"},
 	},
 	{
-		name: "approve",
-		expectedInputs: []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "spender", typeName: "Address"},
-			{name: "amount", typeName: "i128"},
-			{name: "expiration_ledger", typeName: "u32"},
+		Name: "approve",
+		ExpectedInputs: []wasmspec.InputSpec{
+			{Name: "from", TypeName: "Address"},
+			{Name: "spender", TypeName: "Address"},
+			{Name: "amount", TypeName: "i128"},
+			{Name: "expiration_ledger", TypeName: "u32"},
 		},
-		expectedOutputs: []string{},
+		ExpectedOutputs: []string{},
 	},
 	{
-		name: "transfer",
-		expectedInputs: []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-			{name: "amount", typeName: "i128"},
+		Name: "transfer",
+		ExpectedInputs: []wasmspec.InputSpec{
+			{Name: "from", TypeName: "Address"},
+			{Name: "to", TypeName: "Address"},
+			{Name: "amount", TypeName: "i128"},
 		},
 	},
 	// CAP-67 variant: (from: Address, to_muxed: MuxedAddress, amount: i128) -> ()
 	{
-		name: "transfer",
-		expectedInputs: []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to_muxed", typeName: "MuxedAddress"},
-			{name: "amount", typeName: "i128"},
+		Name: "transfer",
+		ExpectedInputs: []wasmspec.InputSpec{
+			{Name: "from", TypeName: "Address"},
+			{Name: "to_muxed", TypeName: "MuxedAddress"},
+			{Name: "amount", TypeName: "i128"},
 		},
 	},
 	{
-		name: "transfer_from",
-		expectedInputs: []contractFunctionInputSpec{
-			{name: "spender", typeName: "Address"},
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-			{name: "amount", typeName: "i128"},
+		Name: "transfer_from",
+		ExpectedInputs: []wasmspec.InputSpec{
+			{Name: "spender", TypeName: "Address"},
+			{Name: "from", TypeName: "Address"},
+			{Name: "to", TypeName: "Address"},
+			{Name: "amount", TypeName: "i128"},
 		},
-		expectedOutputs: []string{},
+		ExpectedOutputs: []string{},
 	},
 	{
-		name: "burn",
-		expectedInputs: []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "amount", typeName: "i128"},
+		Name: "burn",
+		ExpectedInputs: []wasmspec.InputSpec{
+			{Name: "from", TypeName: "Address"},
+			{Name: "amount", TypeName: "i128"},
 		},
-		expectedOutputs: []string{},
+		ExpectedOutputs: []string{},
 	},
 	{
-		name: "burn_from",
-		expectedInputs: []contractFunctionInputSpec{
-			{name: "spender", typeName: "Address"},
-			{name: "from", typeName: "Address"},
-			{name: "amount", typeName: "i128"},
+		Name: "burn_from",
+		ExpectedInputs: []wasmspec.InputSpec{
+			{Name: "spender", TypeName: "Address"},
+			{Name: "from", TypeName: "Address"},
+			{Name: "amount", TypeName: "i128"},
 		},
-		expectedOutputs: []string{},
+		ExpectedOutputs: []string{},
 	},
 }
 
@@ -383,86 +344,5 @@ var requiredFunctions = []functionSpec{
 // exact-match signature. Suitable for unit testing in isolation; production
 // callers reach this via Validator.Validate.
 func matchSEP41Spec(contractSpec []xdr.ScSpecEntry) bool {
-	requiredSpecs := make(map[string][]functionSpec, len(requiredFunctions))
-	for _, spec := range requiredFunctions {
-		requiredSpecs[spec.name] = append(requiredSpecs[spec.name], spec)
-	}
-
-	foundFunctions := set.NewSet[string]()
-
-	for _, spec := range contractSpec {
-		if spec.Kind != xdr.ScSpecEntryKindScSpecEntryFunctionV0 || spec.FunctionV0 == nil {
-			continue
-		}
-
-		function := spec.FunctionV0
-		funcName := string(function.Name)
-
-		expectedSpecs, isRequired := requiredSpecs[funcName]
-		if !isRequired {
-			continue
-		}
-
-		actualInputs := make([]contractFunctionInputSpec, 0, len(function.Inputs))
-		for _, input := range function.Inputs {
-			actualInputs = append(actualInputs, contractFunctionInputSpec{
-				name:     input.Name,
-				typeName: getTypeName(input.Type.Type),
-			})
-		}
-
-		actualOutputs := make([]string, 0, len(function.Outputs))
-		for _, output := range function.Outputs {
-			actualOutputs = append(actualOutputs, getTypeName(output.Type))
-		}
-
-		for _, expectedSpec := range expectedSpecs {
-			if validateFunctionInputsAndOutputs(actualInputs, actualOutputs, expectedSpec.expectedInputs, expectedSpec.expectedOutputs) {
-				foundFunctions.Add(funcName)
-				break
-			}
-		}
-	}
-
-	return foundFunctions.Cardinality() == len(requiredSpecs)
-}
-
-// validateFunctionInputsAndOutputs checks if a function's signature matches
-// the expected SEP-41 specification. It compares ordered input/output slices
-// with exact arity and exact position-by-position matches.
-func validateFunctionInputsAndOutputs(
-	inputs []contractFunctionInputSpec,
-	outputs []string,
-	expectedInputs []contractFunctionInputSpec,
-	expectedOutputs []string,
-) bool {
-	if len(inputs) != len(expectedInputs) {
-		return false
-	}
-
-	for i := range expectedInputs {
-		if inputs[i] != expectedInputs[i] {
-			return false
-		}
-	}
-
-	if len(outputs) != len(expectedOutputs) {
-		return false
-	}
-
-	for i := range expectedOutputs {
-		if outputs[i] != expectedOutputs[i] {
-			return false
-		}
-	}
-	return true
-}
-
-// getTypeName converts an XDR ScSpecType to its human-readable string
-// representation.
-func getTypeName(scType xdr.ScSpecType) string {
-	if name, ok := scSpecTypeNames[scType]; ok {
-		return name
-	}
-	return "unknown"
+	return wasmspec.Match(contractSpec, requiredFunctions)
 }
