+++ internal/services/sep41/validator_test.go
@@ -211,189 +211,6 @@ func TestIsContractCodeSEP41(t *testing.T) {
 	})
 }
 
-func TestValidateFunctionInputsAndOutputs(t *testing.T) {
-	t.Run("returns true for exact match", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		outputs := []string{"i128"}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.True(t, result)
-	})
-
-	t.Run("returns false for too few inputs", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-		}
-		outputs := []string{"i128"}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.False(t, result)
-	})
-
-	t.Run("returns false for too many inputs", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-			{name: "amount", typeName: "i128"},
-		}
-		outputs := []string{"i128"}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.False(t, result)
-	})
-
-	t.Run("returns false for wrong input parameter name", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "sender", typeName: "Address"},
-			{name: "receiver", typeName: "Address"},
-		}
-		outputs := []string{"i128"}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.False(t, result)
-	})
-
-	t.Run("returns false for wrong input parameter type", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "u32"},
-		}
-		outputs := []string{"i128"}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.False(t, result)
-	})
-
-	t.Run("returns false for reordered inputs", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "to", typeName: "Address"},
-			{name: "from", typeName: "Address"},
-		}
-		outputs := []string{"i128"}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.False(t, result)
-	})
-
-	t.Run("returns false for too few outputs", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		outputs := []string{}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.False(t, result)
-	})
-
-	t.Run("returns false for too many outputs", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		outputs := []string{"i128", "u32"}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.False(t, result)
-	})
-
-	t.Run("returns false for wrong output type", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		outputs := []string{"u32"}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.False(t, result)
-	})
-
-	t.Run("returns false for duplicate outputs", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		outputs := []string{"i128", "i128"}
-
-		expectedInputs := []contractFunctionInputSpec{
-			{name: "from", typeName: "Address"},
-			{name: "to", typeName: "Address"},
-		}
-		expectedOutputs := []string{"i128"}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.False(t, result)
-	})
-
-	t.Run("returns true for empty inputs and outputs", func(t *testing.T) {
-		inputs := []contractFunctionInputSpec{}
-		outputs := []string{}
-
-		expectedInputs := []contractFunctionInputSpec{}
-		expectedOutputs := []string{}
-
-		result := validateFunctionInputsAndOutputs(inputs, outputs, expectedInputs, expectedOutputs)
-		assert.True(t, result)
-	})
-}
-
 // wasmTestdataDir returns the absolute path to the shared WASM testdata directory.
 func wasmTestdataDir() string {
 	_, filename, _, _ := runtime.Caller(0)
